// plugins/main/server/lib/generate-alerts/types.ts
export interface Rule {
  id: string;
  level: number;
  description: string;
  groups: string[];
  firedtimes?: number;
  mail?: boolean;
  gdpr?: string[];
  pci_dss?: string[];
  tsc?: string[];
  hipaa?: string[];
  nist_800_53?: string[];
  nist_800_83?: string[];
  gpg13?: string[];
  pci?: string[];
  frequency?: string;
  info?: string;
}

export interface Agent {
  id: string;
  name: string;
  ip: string;
}

export interface Manager {
  name: string;
}

export interface Cluster {
  name: string;
  node: string;
}

export interface Decoder {
  parent?: string;
  name: string;
}

export interface Input {
  type: string;
}

export interface Location {
  lat: number;
  lon: number;
}

export interface Field {
  timestamp: string;
}

export type GeoLocation = {
  country_name: string;
  location: Location;
  region_name: string;
  city_name: string;
};

export interface Process {
  name: string;
  id: number;
  ppid: number;
}

export interface User {
  name: string;
  id: number;
}

export interface Audit {
  process: Process;
  effective_user: User;
  user: User;
  group: User;
}

export interface SysCheck {
  event: string;
  path: string;
  uname_after: string;
  gname_after: string;
  mtime_before: Date;
  mtime_after: Date;
  size_after: number;
  uid_after: string;
  gid_after: string;
  perm_after: string;
  inode_before: number;
  inode_after: number;
  sha1_after: string;
  changed_attributes: string[];
  md5_after: string;
  sha256_after: string;
  tags: string[];
  audit: Audit;
}

type DataKeys =
  | 'integration'
  | 'aws'
  | 'office365'
  | 'gcp'
  | 'audit'
  | 'oscap'
  | 'title'
  | 'file'
  | 'virustotal'
  | 'vulnerability'
  | 'osquery'
  | 'srcip'
  | 'srcuser'
  | 'srcport'
  | 'win'
  | 'dstuser'
  | 'uid'
  | 'euid'
  | 'tty'
  | 'github'
  | 'system_name'
  | 'id'
  | 'protocol'
  | 'url'
  | 'extra_data'
  | 'type'
  | 'status'
  | 'YARA'
  | 'cis';

export type Data = Partial<Record<DataKeys, any>>;

export type PreDecoder = {
  program_name: string;
  timestamp: string;
  hostname?: string;
};

export interface Alert {
  id: string;
  agent: Agent;
  cluster: Cluster;
  data: Data;
  decoder: Decoder;
  location: string;
  manager: Manager;
  rule: Rule;
  timestamp: string;
  predecoder?: PreDecoder;
  fields?: Field;
  full_log?: string;
  GeoLocation?: GeoLocation;
  input?: Input;
  previous_output?: string;
  syscheck?: Partial<SysCheck>;
}